added App.js, App.css and refomated the index.js
