+++ firebase-quiz/src/components/account.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+
+
+
+
+export default Account;
