+++ firebase-quiz/src/components/score.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+
+
+export default score;
